chore: use fileArrayBuffer for backward-compatibility
